// temper/shared-addon-libraries-lib-data-encode/src/types.ts
export interface ControlCharSpec {
  name: string
  length?: number
  decoder?: string
}